refactor: group RockRunner's panels and scene elements into folders
The game's root had grown to a flat list where a panel registration, the
geometry of a drawn edge and the track's own maths all sat side by side
with nothing but a filename to tell them apart.

panel/ takes the three that register rows on the elements panel: the
camera, the rock and the track. elements/ takes what those rows describe:
the rock's surfaces, its outline, and the path's drawn edge.

Left where they are: the scatter's own panel, which belongs with the
scatter feature it configures rather than with the game's general panels,
and the track's path and chunks, which are the world rather than anything
shown in a row.
