+++ src/views/Games/RockRunner/game/useRockRun.ts
@@ -36,11 +36,11 @@ import { createTrackChunkManager } from '../trackChunks'
 import { createScatterAreaManager } from '../scatter/scatterAreas'
 import { createScatterPanel } from '../scatter/scatterPanel'
 import { SCATTER_AREAS } from '../scatter/illustrations'
-import { DEFAULT_RUN_CAMERA, registerCameraElements } from '../cameraPanel'
-import { registerRockElements } from '../rockPanel'
-import { attachRockStroke } from '../rockStroke'
-import { DEFAULT_ROCK_SURFACE, rockSurfaceById } from '../rockSurfaces'
-import { registerTrackElements, createElementVisibilityHandlers } from '../trackPanel'
+import { DEFAULT_RUN_CAMERA, registerCameraElements } from '../panel/cameraPanel'
+import { registerRockElements } from '../panel/rockPanel'
+import { attachRockStroke } from '../elements/rockStroke'
+import { DEFAULT_ROCK_SURFACE, rockSurfaceById } from '../elements/rockSurfaces'
+import { registerTrackElements, createElementVisibilityHandlers } from '../panel/trackPanel'
 import { createLateralFogUniforms } from '../lateralFog'
 import { createDebrisField } from './debris'
 import { stageColorAt } from '../scatter/textureStages'
